requirements.txt
Django>=5.2,<6.0
openai>=1.30.0,<2.0
numpy>=1.26
python-dotenv>=1.0
Pillow>=10.0
